Rename client package to zebedee
